moved SDSortConfig dataclass to sdsort module
